// users-service/src/configuration/InjectionNames.enum.ts
export enum InjectionNames {
  API_KEY = 'API_KEY',
  GRPC_USERS_CLIENT = 'GRPC_USERS_CLIENT',
  CLIENT_GRPC_PROXY = 'CLIENT_GRPC_PROXY',
  CLIENT_TCP_PROXY = 'CLIENT_TCP_PROXY',
  TCP_USERS_CLIENT = 'TCP_USERS_CLIENT',
  GRPC_CONFIG = 'GRPC_CONFIG',
  TCP_CONFIG = 'TCP_CONFIG',
  USE_SWAGGER = 'USE_SWAGGER',
  HASH_ROUNDS = 'HASH_ROUNDS',
  SECRETS_FROM_ENV = 'SECRETS_FROM_ENV',
  PROJECT_NAME = 'PROJECT_NAME',
  CONNECTION_STRING = 'CONNECTION_STRING',
  SWAGGER_CONFIG = 'SWAGGER_CONFIG',
  HEALTH_CHECK_REST_ADDRESS = 'HEALTH_CHECK_REST_ADDRESS',
  HEALTH_CHECK_DOCUMENTATION_ADDRESS = 'HEALTH_CHECK_DOCUMENTATION_ADDRESS',
  USERS_REST_PORT = 'USERS_REST_PORT',
}